Refactoring of Nodes::Binary::Binary to Nodes::Binary
