+++ lib/msfl_visitors/nodes/containment.rb
@@ -1,7 +1,7 @@
-require_relative 'binary/binary'
+require_relative 'binary'
 module MSFLVisitors
   module Nodes
-    class Containment < Binary::Binary
+    class Containment < Binary
     end
   end
 end
